Changes to forms, mixins, filters, and rehal of thing detail

+++ core/models/comment.py
@@ -6,7 +6,7 @@ from django.urls import reverse
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.utils import timezone
 
-from django.db.models.fields import BooleanField, CharField, DateField, DecimalField, TextField
+from django.db.models.fields import BooleanField, CharField, DateField, SmallIntegerField, TextField
 from django.db.models.fields.related import ForeignKey
 
 from .thing import Thing
@@ -15,10 +15,7 @@ from .user import UserProfile
 class Comment(models.Model):
     title = CharField(max_length=128)
     content = TextField(max_length=4096)
-    rating = DecimalField(validators=[MaxValueValidator(limit_value=5), MinValueValidator(limit_value=1)],
-        max_digits=2,
-        decimal_places=1)
-    
+    rating = SmallIntegerField(validators=[MaxValueValidator(limit_value=5), MinValueValidator(limit_value=1)])    
     posted_on = DateField(default=timezone.now)
     is_edited = BooleanField(default=False)
